refactor(server): extract routes function to controller
